pyproject.toml
[tool.poetry]
name = "check_mkdocs"
version = "0.0.1"
description = ""
readme = "README.md"
authors = ["Rodrigo Gonzalez <r@rodrigo-gonzalez.com>"]
repository = "https://github.com/RodrigoGonzalez/check-mkdocs"
license = "MIT"

packages = [
    {include = "check_mkdocs", from = "src"}
]
include = ["README.rst", "LICENSE", "CHANGELOG.md"]

[tool.poetry.scripts]
import-rules = "check_mkdocs.__main__:main"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.12.0"
commitizen = "^3.0.1"
pre-commit = "^3.2.2"
isort = "^5.12.0"
black = "^23.3.0"
autoflake = "^2.2.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
coverage = { version = "^7.3.0",  extras = ["toml"] }
covdefaults = "^2.3.0"
pytest-mock = "^3.11.1"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.0"
mkdocs-material = "^9.1.6"
mkdocstrings = {extras = ["optimize", "python"], version = "^0.22.0"}

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.1"
tag_format = "v$version"
version_type = "semver"
update_changelog_on_bump = true
version_files = [
    "pyproject.toml:version",
    "src/check_mkdocs/__init__.py:__version__",
]
changelog_file = "CHANGELOG.md"

[tool.isort]
profile = "black"
skip_glob = ["docs", "example_repos"]
line_length = 100
multi_line_output = 4
force_single_line = true
src_paths = ["src", "tests", "scripts"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_first_party = ["check_mkdocs"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
    /(
        \.git
      | \.mypy_cache
      | \.mypy_html_report
      | \.venv
      | _build
      | build
      | dist
      | docs
      | example_repos
    )/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
